ISPN-13156 Multiple instances of XMLTranscoder reuse the same internal object
